Re-factored out some circular deps after call-by-name
